+++ test/AuctionHouse/settle.t.sol
@@ -7,7 +7,7 @@ import {Auctioneer} from "src/bases/Auctioneer.sol";
 import {AuctionHouseTest} from "test/AuctionHouse/AuctionHouseTest.sol";
 
 contract SettleTest is AuctionHouseTest {
-    uint96 internal constant _BID_AMOUNT_TOTAL = 20e18;
+    uint256 internal constant _BID_AMOUNT_TOTAL = 20e18;
 
     uint256 internal _expectedSellerQuoteTokenBalance;
     uint256 internal _expectedBidderQuoteTokenBalance;
@@ -118,15 +118,15 @@ contract SettleTest is AuctionHouseTest {
     }
 
     modifier givenLotHasPartialFill() {
-        uint96 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
-        uint96 totalOut = _scaleBaseTokenAmount(_LOT_CAPACITY);
+        uint256 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
+        uint256 totalOut = _scaleBaseTokenAmount(_LOT_CAPACITY);
         uint256 scaledLotCapacity = _scaleBaseTokenAmount(_LOT_CAPACITY);
 
         // Total bid was 4e18, since 2e18 of quote token = 1e18 of base token
-        uint96 pfRefundAmount = _scaleQuoteTokenAmount(2e18);
-        uint96 pfPayoutAmount = _scaleBaseTokenAmount(1e18);
-        uint96 pfFilledAmount = _scaleQuoteTokenAmount(4e18) - pfRefundAmount;
-        uint96 totalInFilled = totalIn - pfRefundAmount;
+        uint256 pfRefundAmount = _scaleQuoteTokenAmount(2e18);
+        uint256 pfPayoutAmount = _scaleBaseTokenAmount(1e18);
+        uint256 pfFilledAmount = _scaleQuoteTokenAmount(4e18) - pfRefundAmount;
+        uint256 totalInFilled = totalIn - pfRefundAmount;
 
         Auction.Settlement memory settlement = Auction.Settlement({
             totalIn: totalIn,
@@ -179,8 +179,8 @@ contract SettleTest is AuctionHouseTest {
     }
 
     modifier givenLotIsUnderCapacity() {
-        uint96 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
-        uint96 totalOut = _scaleBaseTokenAmount(5e18); // 50% filled
+        uint256 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
+        uint256 totalOut = _scaleBaseTokenAmount(5e18); // 50% filled
         uint256 scaledLotCapacity = _scaleBaseTokenAmount(_LOT_CAPACITY);
 
         Auction.Settlement memory settlement = Auction.Settlement({
@@ -234,8 +234,8 @@ contract SettleTest is AuctionHouseTest {
     }
 
     modifier givenLotCapacityIsFilled() {
-        uint96 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
-        uint96 totalOut = _scaleBaseTokenAmount(_LOT_CAPACITY);
+        uint256 totalIn = _scaleQuoteTokenAmount(_BID_AMOUNT_TOTAL);
+        uint256 totalOut = _scaleBaseTokenAmount(_LOT_CAPACITY);
         uint256 scaledLotCapacity = _scaleBaseTokenAmount(_LOT_CAPACITY);
 
         Auction.Settlement memory settlement = Auction.Settlement({
@@ -289,8 +289,8 @@ contract SettleTest is AuctionHouseTest {
     }
 
     modifier givenLotDoesNotSettle() {
-        uint96 totalIn = 0;
-        uint96 totalOut = 0;
+        uint256 totalIn = 0;
+        uint256 totalOut = 0;
         uint256 scaledLotCapacity = _scaleBaseTokenAmount(_LOT_CAPACITY);
 
         Auction.Settlement memory settlement = Auction.Settlement({
